requirements.txt
streamlit
python-dotenv
google-cloud-aiplatform
vertexai
google-genai
langchain
langchain-google-vertexai
langchain-core
pydantic